requirements.txt
mem0ai
qdrant-client
